requirements.txt
# Core AgentBeats and A2A dependencies
a2a-sdk>=0.3.5
google-adk>=1.14.1
google-genai>=1.36.0
pydantic>=2.11.9
python-dotenv>=1.1.1
uvicorn>=0.35.0

# Bargaining environment dependencies
numpy>=1.26.0
cvxpy>=1.4.0

# Note: The agentbeats package is included as a local package in src/agentbeats/
# Python buildpacks will automatically include it when deploying from the repo root.

